navbar completed & committing all files
